move BaDataProviders and BaUiProviders components out of app as they are not routes
